Refactor for SOLID
