+++ src/main/java/com/capgemini/wsb/fitnesstracker/configs/ApplicationConfiguration.java
@@ -1,9 +1,17 @@
 package com.capgemini.wsb.fitnesstracker.configs;
 
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.JavaMailSenderImpl;
 
 @Configuration
 @EnableAspectJAutoProxy
 public class ApplicationConfiguration {
+
+    @Bean
+    public JavaMailSender javaMailSender() {
+        return new JavaMailSenderImpl();
+    }
 }
