+++ src/main/java/com/capgemini/wsb/fitnesstracker/statistics/internal/StatisticsMapper.java
@@ -2,14 +2,10 @@ package com.capgemini.wsb.fitnesstracker.statistics.internal;
 
 import com.capgemini.wsb.fitnesstracker.statistics.api.Statistics;
 import com.capgemini.wsb.fitnesstracker.statistics.api.StatisticsDto;
-import com.capgemini.wsb.fitnesstracker.training.api.Training;
-import com.capgemini.wsb.fitnesstracker.training.api.TrainingDto;
-import com.capgemini.wsb.fitnesstracker.training.api.TrainingInputDto;
-import com.capgemini.wsb.fitnesstracker.user.api.User;
 import org.springframework.stereotype.Component;
 
 @Component
-public class StatisticsMapper {
+class StatisticsMapper {
 
     StatisticsDto toDto(Statistics statistics) {
         return new StatisticsDto(statistics.getId(),
