+++ src/main/java/com/capgemini/wsb/fitnesstracker/training/api/Training.java
@@ -1,6 +1,5 @@
 package com.capgemini.wsb.fitnesstracker.training.api;
 
-import com.capgemini.wsb.fitnesstracker.training.internal.ActivityType;
 import com.capgemini.wsb.fitnesstracker.user.api.User;
 import jakarta.persistence.*;
 import lombok.*;
