+++ src/main/java/com/capgemini/wsb/fitnesstracker/training/api/TrainingDto.java
@@ -1,6 +1,5 @@
 package com.capgemini.wsb.fitnesstracker.training.api;
 
-import com.capgemini.wsb.fitnesstracker.training.internal.ActivityType;
 import com.capgemini.wsb.fitnesstracker.user.api.User;
 import jakarta.annotation.Nullable;
 import lombok.Getter;
