+++ src/main/java/com/capgemini/wsb/fitnesstracker/training/api/TrainingInputDto.java
@@ -1,6 +1,5 @@
 package com.capgemini.wsb.fitnesstracker.training.api;
 
-import com.capgemini.wsb.fitnesstracker.training.internal.ActivityType;
 import lombok.Getter;
 import lombok.Setter;
