+++ src/main/java/com/capgemini/wsb/fitnesstracker/training/api/TrainingProvider.java
@@ -1,10 +1,7 @@
 package com.capgemini.wsb.fitnesstracker.training.api;
 
-import com.capgemini.wsb.fitnesstracker.training.internal.ActivityType;
-
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 public interface TrainingProvider {
